core: relocate live/replay tracker impl headers into cpp/inc/<module>/
The live_trackers and replay_trackers modules were the only two src/core
modules keeping implementation headers (*_impl.hpp) loose in their bare
cpp/ directory; every other module places all headers under
cpp/inc/<module>/ (the module's INTERFACE-exported include root). This
relocates the 13 impl headers (7 live, 6 replay) into cpp/inc/<module>/
beside the public headers and fixes the include paths and CMake source
lists to match. Pure file move + include/CMake path fixups — no
behaviour change (build green incl. clang_format_check; ctest 139/139).

- git mv the 13 *_impl.hpp into cpp/inc/<module>/.
- Reprefix in-module includes to "inc/<module>/...": the 13 impl .cpp
  self-includes and the two *_deviceio_factory.cpp dispatch blocks.
- The two moved live impl headers that included the schema header now
  include it as a bare sibling ("schema_tracker.hpp"), since they share
  its directory after the move.
- Repath (not delete) the .hpp entries in both cpp/CMakeLists.txt;
  target_include_directories / link lists are unchanged.
- Document the public/private header convention in
  live_trackers/AGENTS.md and a new replay_trackers/AGENTS.md, and point
  the "New tracker" guidance at the new header location.

Addresses TR-07. Part of #583.
