+++ src/core/live_trackers/cpp/CMakeLists.txt
@@ -16,13 +16,13 @@ add_library(live_trackers STATIC
     inc/live_trackers/schema_tracker_base.hpp
     inc/live_trackers/schema_tracker.hpp
     inc/live_trackers/live_deviceio_factory.hpp
-    live_head_tracker_impl.hpp
-    live_hand_tracker_impl.hpp
-    live_controller_tracker_impl.hpp
-    live_message_channel_tracker_impl.hpp
-    live_full_body_tracker_pico_impl.hpp
-    live_generic_3axis_pedal_tracker_impl.hpp
-    live_frame_metadata_tracker_oak_impl.hpp
+    inc/live_trackers/live_head_tracker_impl.hpp
+    inc/live_trackers/live_hand_tracker_impl.hpp
+    inc/live_trackers/live_controller_tracker_impl.hpp
+    inc/live_trackers/live_message_channel_tracker_impl.hpp
+    inc/live_trackers/live_full_body_tracker_pico_impl.hpp
+    inc/live_trackers/live_generic_3axis_pedal_tracker_impl.hpp
+    inc/live_trackers/live_frame_metadata_tracker_oak_impl.hpp
 )
 
 target_include_directories(live_trackers
