+++ src/core/replay_trackers/cpp/replay_generic_3axis_pedal_tracker_impl.cpp
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-#include "replay_generic_3axis_pedal_tracker_impl.hpp"
+#include "inc/replay_trackers/replay_generic_3axis_pedal_tracker_impl.hpp"
 
 #include <mcap/recording_traits.hpp>
 #include <schema/pedals_bfbs_generated.h>
